Initial implementation of P2719
Squashed version of https://github.com/llvm/llvm-project/pull/113510 as
of November 25 (bc51d9ce1c513b3ef44ad4ed47b2e3ad6077bd3a).

This is a basic implementation of P2719: "Type-aware allocation and
deallocation functions" described at http://wg21.link/P2719

The proposal includes some more details but the basic change in functionality
is the addition of support for an additional implicit parameter in operators
`new` and `delete` to act as a type tag. Tag is of type `std::type_identity<T>`
where T is the concrete type being allocated. So for example, a custom type
specific allocator for `int` say can be provided by the declaration of

  void *operator new(std::type_identity<int>, size_t);
  void  operator delete(std::type_identity<int>, void*);

However this becomes more powerful by specifying templated declarations, for
example

  template <typename T> void *operator new(std::type_identity<T>, size_t);
  template <typename T> void  operator delete(std::type_identity<T>, void*);

Where the operators being resolved will be the concrete type being operated
over (NB. A completely unconstrained global definition as above is not
recommended as it triggers many problems similar to a general override of
the global operators).

These type aware operators can be declared as either free functions or in
class, and can be specified with or without the other implicit parameters,
with overload resolution performed according to the existing standard parameter
prioritisation, only with type parameterised operators having higher precedence
than non-type aware operators. The only exception is destroying_delete which
for reasons discussed in the paper we do not support type-aware variants by
default.

In order to ensure interchangability with existing operators, this implementation
does extend the definition of a `usual deallocation function` as suggested in the
proposal to support templated operators, as long as the only dependent parameter
is the type_identity tag type, and the tag is explicitly a specialization of
std::type_identity

The implementation adds a new `-fexperimental-cxx-type-aware-allocators` flag
to enable the feature. This is detectable in source via the `cxx_type_aware_allocators`
feature check.

This also adds a separate flag to support the application of the proposal
to the destroying delete, `-fexperimental-cxx-type-aware-destroying-delete`.
The proposal currently does not support destroying delete as it adds some
significant hazards, but we've received sufficient requests for the
behavior that we included it so people can see if it is actually useful.
This can be detected in source via the `cxx_type_aware_destroying_delete`
feature flag.

The implementation includes some error checking and warnings to try to
detect basic errors, but the problem of detecting mismatch type-polymorphic
new and delete is a problem to be explored in future.

Update clang/lib/AST/DeclCXX.cpp

Co-authored-by: Erich Keane <ekeane@nvidia.com>

Update clang/lib/AST/Type.cpp

Co-authored-by: Aaron Ballman <aaron@aaronballman.com>

Update clang/include/clang/Basic/DiagnosticSemaKinds.td

Co-authored-by: Aaron Ballman <aaron@aaronballman.com>

Addressing review comments

Yet more review comments

Minor edits

Missed a -fexperimental change and adding flag names to diagnostics

Correct formatting

Sigh, let's appease MSVC's refusal to acknowledge designated initializers

Renaming this function to be more clear that it is not just a lang opt

Add a test for qualifiers out of a sense of paranoia

Add an explicit test to ensure type aware delete is called for cleanup

Annoyingly this needs to be a codegen test for sema. The long term fix is
probably to make new expressions actually build up a full AST node of
dependent expressions, rather than including a bunch of flags to tell
codegen what to do.

More paranoia tests

Reformat following the grand enum refactoring

Update error messages, and subsequently the tests

Reduce the places we pass the implicit parameters separately

Added codegen tests

This did find a few places where we were not forcing full instantiation
of the type_identity specialization when calling the clean up method in
an environment where the relevant new and delete operators do not have
matching type-awareness.

Formatting remains my nemesis

Increase test coverage and validation

Current P2719 does not make this an error, so leave as a warning pending language approval

select correct base for multidimensional arrays